fix: extract board service as library
