--- apps/api/src/utils/array.ts
@@ -1,3 +1,0 @@
-export function emptyBoard(): number[][] {
-  return new Array(6).fill([]).map(() => new Array(6).fill(0));
-}
