+++ libs/board/src/board.model.ts
@@ -28,3 +28,7 @@ export class Board extends Model {
   @Column
   hash: string;
 }
+
+export function emptyBoard(): number[][] {
+  return new Array(6).fill([]).map(() => new Array(6).fill(0));
+}
