+++ libs/board/src/index.ts
@@ -1,0 +1,2 @@
+export * from './board.module';
+export * from './board.service';
