+++ tsconfig.json
@@ -17,6 +17,13 @@
     "strictBindCallApply": true,
     "forceConsistentCasingInFileNames": true,
     "noFallthroughCasesInSwitch": true,
-    "paths": {}
+    "paths": {
+      "@board/board": [
+        "libs/board/src"
+      ],
+      "@board/board/*": [
+        "libs/board/src/*"
+      ]
+    }
   }
 }
